Replace calls to Math.pow(10, x) with constants
There are a number of places where a fixed calculation of Math.pow is
used. Typically to determine units so 10^3, 10^6, 10^9 etc. These are
being recalculated on every format call. Since formatting could
potentially be triggered quite frequently, I felt it made sense to
optimize these as constants for 3, 6, 9 and 12 which are the only
usages.